% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_budget.R
\name{clean_extract_PBS_xlsx}
\alias{clean_extract_PBS_xlsx}
\title{This function will extract out data from PBS Budget data from data.gov}
\usage{
clean_extract_PBS_xlsx(
  .path = "2021-22 PBS AFP.XLSX",
  entity_name = "Australian Federal Police",
  budget_calendar_year = 2021
)
}
\arguments{
\item{.path}{(character) Path to excel file containing the PBS data.}

\item{entity_name}{(character) The entity name of the data in the excel file}

\item{budget_calendar_year}{(numeric) The calendar year the budget was
published.}
}
\value{
(list) Returns a list of the tables extracted from the sheet.
}
\description{
This function will extract out data from PBS Budget data from data.gov
}
\examples{
\dontrun{

clean_extract_PBS_xlsx(
.path = "2021-22 PBS AFP.XLSX",
 entity_name = "Australian Federal Police",
 budget_calendar_year = 2021)

}
}
