% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_CED_concord.R
\name{concord_CED_POA_by_pop}
\alias{concord_CED_POA_by_pop}
\title{concord_CED_POA_by_pop}
\usage{
concord_CED_POA_by_pop(
  MB_pop = get_ABS_MB_pop(),
  POA_data = get_abs_region_allocation(geo_type = "POA"),
  CED_data = get_abs_region_allocation(geo_type = "CED")
)
}
\arguments{
\item{MB_pop}{(tibble) Meshblock Population. By Default it will draw this from the ABS
website using an internal function in this package.}

\item{POA_data}{(tibble) POA Allocation Data. By Default it will draw this from the ABS
website using an internal function in this package.}

\item{CED_data}{(tibble) CED Allocation Data. By Default it will draw this from the ABS
website using an internal function in this package.}
}
\value{
(tibble) Returns a tibble with POA and CED concorded with percentage allocation
based on Meshblock population
}
\description{
This function will conocord CED to POA using meshblock population.
All allocation data is downloaded from:
"https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/allocation-files/"
}
