% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{convert_month}
\alias{convert_month}
\title{Detect a string and converts that string from the month abbreviation into
a number}
\usage{
convert_month(.vec)
}
\arguments{
\item{.vec}{(character) Input character vector}
}
\description{
Detect a string and converts that string from the month abbreviation into
a number
}
\examples{
\dontrun{

strings <- c("I am January", "You 03 February test TEst")

extracted_month <- convert_month(strings)

}
}
