% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{days_of_week}
\alias{days_of_week}
\title{Return a vector of days of the week}
\usage{
days_of_week(abbrev = FALSE)
}
\arguments{
\item{abbrev}{(Boolean; FALSE) Abbreviates the days if set to TRUE}
}
\description{
Return a vector of days of the week
}
\examples{
\dontrun{
days_week <- days_of_week()
}
}
