% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{download_read_xlsx}
\alias{download_read_xlsx}
\title{Download and read in xlsx file}
\usage{
download_read_xlsx(url)
}
\arguments{
\item{url}{(character) The download URL for the xlsx file}
}
\value{
(tibble)
}
\description{
Download and read in xlsx file
}
\examples{
\dontrun{
url <- "https://data.gov.au/data/dataset/863c394c-a26c-4340-896b-a26b18af476d/resource/66ee77e9-34b0-4215-8bc3-bfa43b7c8f11/download/2018-19-pbs-program-expense-line-items-1.xlsx"
data <- download_xlsx(url)
}
}
