% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_budget.R
\name{extract_aus_BP2}
\alias{extract_aus_BP2}
\title{Extract out AUS Budget Paper 2 Data from word doc}
\usage{
extract_aus_BP2(skip_tables = 32, financial_year = 2022)
}
\arguments{
\item{financial_year}{}
}
\description{
Extract out AUS Budget Paper 2 Data from word doc
}
