% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_tender.R
\name{extract_single_abn_data}
\alias{extract_single_abn_data}
\title{Go to the ABN website and extract information about a single ABN}
\usage{
extract_single_abn_data(abn = "42000001007")
}
\arguments{
\item{abn}{(character or numeric) The ABN you are looking to extract information
from.}
}
\value{
(tibble) Returns a tibble of extracted information.
}
\description{
Go to the ABN website and extract information about a single ABN
}
\examples{
\dontrun{

abns <- c("50128382187", "42000001007")

data <- abns \%>\%
 purrr::map_dfr(extract_single_abn_data)

}
}
