% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs.R
\name{get_abs_cpi_table1}
\alias{get_abs_cpi_table1}
\title{Get CPI TABLES 1 and 2. CPI: All Groups, Index Numbers and Percentage Changes
and clean}
\usage{
get_abs_cpi_table1()
}
\description{
Get CPI TABLES 1 and 2. CPI: All Groups, Index Numbers and Percentage Changes
and clean
}
