% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_CED_concord.R
\name{get_ABS_MB_pop}
\alias{get_ABS_MB_pop}
\title{get_ABS_MB_pop}
\usage{
get_ABS_MB_pop(
  url_x =
    "https://www.abs.gov.au/census/guide-census-data/mesh-block-counts/2021/Mesh\%20Block\%20Counts\%2C\%202021.xlsx"
)
}
\arguments{
\item{url_x}{}
}
\description{
This function will fetch the ABS population by meshblock from the ABS website
}
\details{
Data is downloaded from:
"https://www.abs.gov.au/census/guide-census-data/mesh-block-counts/2021/"
}
