% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs.R
\name{get_abs_quick_stats}
\alias{get_abs_quick_stats}
\title{Download ABS quickstats with your chosen year, geography and
geo code. The geo code corrosponds to the code for the
region you are interested in ie; 10180 for Armadale Region LGA.}
\usage{
get_abs_quick_stats(year = "2021", geography = "LGA", geo_code = "10180")
}
\arguments{
\item{year}{(character; 2021) The year of data you want}

\item{geography}{(character; "LGA") Geography type, ie; SA2, SA3
LGA, etc}

\item{geo_code}{(character; "10180") The code corrosponding to
the region you want ie; "10180" for Armadale Region, and
"116011303" for Blacktown east SA2 etc.
get region files here:
https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/allocation-files}
}
\value{
(tibble)
}
\description{
Download ABS quickstats with your chosen year, geography and
geo code. The geo code corrosponds to the code for the
region you are interested in ie; 10180 for Armadale Region LGA.
}
\examples{
\dontrun{

dat <- get_abs_quick_stats(
year = "2016",
geography = "LGA",
geo_code = "10180"
)



}
}
