% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs.R
\name{get_abs_region_allocation}
\alias{get_abs_region_allocation}
\title{This function plucks concordance files from the ABS website and reads it into
your global environment.}
\usage{
get_abs_region_allocation(geo_type = "SA2")
}
\arguments{
\item{geo_type}{(character; SA2) Geo type, set to POA for postcodes and LGA
for LGA.}
}
\value{
(tibble)
}
\description{
All allocation data is downloaded from:
"https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/allocation-files/"
}
\examples{
\dontrun{

#POA data
post_code_concordances <- get_abs_region_allocation("POA")

#State Electorate Data

POA_data <-
 get_abs_region_allocation(geo_type = "POA") \%>\%
 dplyr::distinct(MB_CODE_2021, POA_CODE_2021)

#State Electorate Data

CED_data <-
 get_abs_region_allocation(geo_type = "CED") \%>\%
 dplyr::distinct(MB_CODE_2021, CED_CODE_2021, CED_NAME_2021, STATE_NAME_2021)

#State Electorate Data
SED_data <-
 get_abs_region_allocation(geo_type = "SED") \%>\%
 dplyr::distinct(MB_CODE_2021, SED_CODE_2024, SED_NAME_2024, STATE_NAME_2021)

}
}
