% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoL_functions.R
\name{get_bls_2000_to_now}
\alias{get_bls_2000_to_now}
\title{This function will retrieve stack multiple v1 queries together to
get data starting from the year 2000. This is required because of the
limited nature of the v1 API. Querying any 'popular series'
is limited to 10 years using the v1 API.}
\usage{
get_bls_2000_to_now(series)
}
\arguments{
\item{series}{(character) The series ID of the data you wish to retrieve.
https://data.bls.gov/cgi-bin/surveymost?bls to find out what seriesid's are
available for use with this function and what data sets they retrieve from BLS.}
}
\value{
(tibble) Returns a data frame
}
\description{
This function will retrieve stack multiple v1 queries together to
get data starting from the year 2000. This is required because of the
limited nature of the v1 API. Querying any 'popular series'
is limited to 10 years using the v1 API.
}
\examples{
\dontrun{

get_bls_2000_to_now(series = 'CES0000000001')

}
}
