% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoL_functions.R
\name{get_bls_total_avg_wkly_hours}
\alias{get_bls_total_avg_wkly_hours}
\title{This function will retrieve Private average weekly hours Seasonally adjusted
data from BLS starting from 2000 to the current year.}
\usage{
get_bls_total_avg_wkly_hours()
}
\value{
(tibble) Returns a data frame
}
\description{
This function will retrieve Private average weekly hours Seasonally adjusted
data from BLS starting from 2000 to the current year.
}
\examples{
\dontrun{

dat <- get_bls_total_avg_wkly_hours()

}
}
