% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoL_functions.R
\name{get_bls_total_avg_wkly_hours_of_prod}
\alias{get_bls_total_avg_wkly_hours_of_prod}
\title{This function will retrieve Total Private Average Weekly Hours of Prod.
and Nonsup. Employees - Seasonally Adjusted - CES0500000007 data from BLS
starting from 2000 to the current year.}
\usage{
get_bls_total_avg_wkly_hours_of_prod()
}
\value{
(tibble) Returns a data frame
}
\description{
This function will retrieve Total Private Average Weekly Hours of Prod.
and Nonsup. Employees - Seasonally Adjusted - CES0500000007 data from BLS
starting from 2000 to the current year.
}
\examples{
\dontrun{

dat <- get_bls_total_LF()

}
}
