% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoL_functions.R
\name{get_bls_total_avg_wkly_hrly_earnings}
\alias{get_bls_total_avg_wkly_hrly_earnings}
\title{This function will retrieve Total Private Average Hourly Earnings of All Employees
\itemize{
\item Seasonally Adjusted - CES0500000003 data from BLS starting from 2000 to
the current year.
}}
\usage{
get_bls_total_avg_wkly_hrly_earnings()
}
\value{
(tibble) Returns a data frame
}
\description{
This function will retrieve Total Private Average Hourly Earnings of All Employees
\itemize{
\item Seasonally Adjusted - CES0500000003 data from BLS starting from 2000 to
the current year.
}
}
\examples{
\dontrun{

dat <- get_bls_total_avg_wkly_hrly_earnings()

}
}
