% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoL_functions.R
\name{get_bls_total_civ_emp}
\alias{get_bls_total_civ_emp}
\title{This function will retrieve Civilian Employment Seasonally adjusted data from BLS
starting from 2000 to the current year.}
\usage{
get_bls_total_civ_emp()
}
\value{
(tibble) Returns a data frame
}
\description{
This function will retrieve Civilian Employment Seasonally adjusted data from BLS
starting from 2000 to the current year.
}
\examples{
\dontrun{

dat <- get_bls_total_civ_emp()

}
}
