% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_CED_concord.R
\name{get_CED_redist}
\alias{get_CED_redist}
\title{get_WA_CED_redist}
\usage{
get_CED_redist(
  MB_allocation_file = get_abs_region_allocation(),
  region_allocation = "NSW electorate"
)
}
\value{
(tibble) Tibble of Redistribution
}
\description{
get_WA_CED_redist
}
\examples{
\dontrun{

WA_redist <- get_WA_CED_redist()

}
}
