% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_fx_data.R
\name{get_confidence_events_fx_calender}
\alias{get_confidence_events_fx_calender}
\title{Extracts confidence data from \code{get_macro_event_data()}, this includes things
like Australian Westpac Confidence, Australian AIG indexes}
\usage{
get_confidence_events_fx_calender(
  .data = get_macro_event_data(),
  column_affix = "conf",
  country_symbol = NULL,
  wide = FALSE
)
}
\arguments{
\item{.data}{(tibble;get_macro_event_data()) Daily fx macro economic calendar
data}

\item{column_affix}{(character;"conf") This is what is added to the variables
to signify that they are confidence measures.}

\item{country_symbol}{(character; NULL) filters for a currency symbol in the
data. Ie; "AUD", "CHF", etc.}

\item{wide}{(Boolean; FALSE) Pivots the data wider.}
}
\value{
(tibble)
}
\description{
Extracts confidence data from \code{get_macro_event_data()}, this includes things
like Australian Westpac Confidence, Australian AIG indexes
}
