% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\name{get_geo_pol_events}
\alias{get_geo_pol_events}
\title{Extracts the table of geopolitical events from Wikipedia.}
\usage{
get_geo_pol_events()
}
\value{
(tibble)
}
\description{
Extracts the table of geopolitical events from Wikipedia.
}
\examples{
\dontrun{

geo_events <- get_geo_pol_events()

}
}
