% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_fx_data.R
\name{get_macro_event_data}
\alias{get_macro_event_data}
\title{Returns data scraped from Macro economic calender releases from dailyfx.com
The exact sources for each of the reported data can be found from the dailyfx
macroeconomic calenders. This includes lots of releases from GDP, to PMI
and other various sentiment indicators from various countries.}
\usage{
get_macro_event_data()
}
\value{
(tibble) A tibble of economic data releases for various countries.
Includes
}
\description{
Returns data scraped from Macro economic calender releases from dailyfx.com
The exact sources for each of the reported data can be found from the dailyfx
macroeconomic calenders. This includes lots of releases from GDP, to PMI
and other various sentiment indicators from various countries.
}
\examples{
\dontrun{

macro_data <- get_macro_event_data()

}
}
