% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_finance.R
\name{get_market_watch}
\alias{get_market_watch}
\title{This function will extract data from Market Watch  based on the asset symbol
provided. The user specifies the start date and end date either as a character
or a date variable type. The function will return data for the asset between
the date ranges.}
\usage{
get_market_watch(
  asset = "tmbmkgb-10y",
  start_date = lubridate::today() - lubridate::years(1),
  end_date = lubridate::today()
)
}
\arguments{
\item{asset}{(character; "GBPUSD") Character value for the symbol of the
asset the user wishes to exrtact.}

\item{start_date}{(Character or date; "2008-01-01")}

\item{end_date}{(Character or date; lubridate::today())}
}
\value{
(tibble) Returns data extracted from market watch.
}
\description{
This function will extract data from Market Watch  based on the asset symbol
provided. The user specifies the start date and end date either as a character
or a date variable type. The function will return data for the asset between
the date ranges.
}
\examples{
\dontrun{

asset <- "AUDUSD"
dat <- get_market_watch(asset = "AUDUSD")

}
}
