% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obr.R
\name{get_obr_public_finances}
\alias{get_obr_public_finances}
\title{This function will go to the OBR website and attempt to the the latest data kept
in the public finances since 1900 sheet. These values are expressed as a \% of
GDP.
OBR: Office of budget responsibility, a publically funded organisation of
the UK government.}
\usage{
get_obr_public_finances()
}
\value{
(tibble) Returns a tibble containing Total Receipts as \% of GDP,
Total Spending as \% of GDP, Public sector net borrowing, Public sector net debt
}
\description{
This function will go to the OBR website and attempt to the the latest data kept
in the public finances since 1900 sheet. These values are expressed as a \% of
GDP.
OBR: Office of budget responsibility, a publically funded organisation of
the UK government.
}
\examples{
\dontrun{
dat <- get_obr_public_finances()
}
}
