% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_data_helpers.R
\name{get_oecd_avlble_indctors_DPLIVE}
\alias{get_oecd_avlble_indctors_DPLIVE}
\title{This function will look inside the 'DP_LIVE' data catalog and return a tibble
of indicators that are available inside of the OECD database. 'DP_LIVE'
contains various different economic, social and government indicators
reported by countries to the OECD. These are not harmonized statistics,
so when looking at things like GDP, CPI you will be viewing the data as
reported by the Country. These figures should align with reported numbers
on various FX calender websites or the relevant statistics agencies for the
relevant country.}
\usage{
get_oecd_avlble_indctors_DPLIVE()
}
\value{
(tibble)
}
\description{
This function will look inside the 'DP_LIVE' data catalog and return a tibble
of indicators that are available inside of the OECD database. 'DP_LIVE'
contains various different economic, social and government indicators
reported by countries to the OECD. These are not harmonized statistics,
so when looking at things like GDP, CPI you will be viewing the data as
reported by the Country. These figures should align with reported numbers
on various FX calender websites or the relevant statistics agencies for the
relevant country.
}
\examples{
\dontrun{

available_indicators <- get_oecd_avlble_indctors_DPLIVE()

GDP_based_indicators <- available_indicators \%>\%
   filter(str_detect(indicator_desc, "GDP|gdp")) \%>\%
   pull(INDICATOR)

dat <- get_oecd_data_DP_LIVE(
  indicators = GDP_based_indicators,
   countries = c("USA"),
   start_time = 2020,
   end_time = 2021
)

}
}
