% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_data_helpers.R
\name{get_oecd_country_abbrevs}
\alias{get_oecd_country_abbrevs}
\title{This function will return a tibble containg country and names and abbrevations
used in the OECD database. The other OECD get functions require country
abbreviations to be provided as an input.}
\usage{
get_oecd_country_abbrevs()
}
\value{
(tibble) A tibble containing all country abbreviations and names.
}
\description{
This function will return a tibble containg country and names and abbrevations
used in the OECD database. The other OECD get functions require country
abbreviations to be provided as an input.
}
\examples{
\dontrun{

country_abbrevs <- get_oecd_country_abbrevs() \%>\%
 pull(.data$id)

GDP_dat <-
  get_oecd_gdp(
    variables = "QNA",
    countries = country_abbrevs,
    start_time = 2006,
    endtime = 2022,
    remove_abbrev_cols = TRUE
  )

}
}
