% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_data_helpers.R
\name{get_oecd_gdp_live}
\alias{get_oecd_gdp_live}
\title{This function will use the package \code{OECD} to extract GDP data for select
countries. This returns a tibble of all detailed GDP data which has been
harmonized the OECD for allow for country to country comparisons.}
\usage{
get_oecd_gdp_live(
  variables = "QNA",
  countries = c("CHN", "USA", "GBR", "OECD", "G-7", "EU27_2020", "JPN", "ITA", "AUS",
    "CAN", "FRA", "DEU", "NZL", "NLD", "KOR", "IDN", "CHE"),
  start_time = 2006,
  end_time = 2022,
  remove_abbrev_cols = TRUE
)
}
\arguments{
\item{variables}{(character; "QNA") This variables which database table the
function will extract data from. The default is "QNA" which is the quarterly
national account table.}

\item{countries}{(character) Country abbreviations used by the function to
determine which countries will be extracted. Use \code{get_oecd_country_abbrevs}
to get the different abbreviations and associated country names.}

\item{start_time}{(numeric) The earliest year you want to go back to. The
function will attempt to go back as far as it can for the countries chosen.}

\item{end_time}{(numeric) Latest date you wish to extract data for.}

\item{remove_abbrev_cols}{(Boolean; TRUE) Will remove unncessary abbreviation
columns if set to TRUE.}
}
\value{
(tibble) Tibble containing the GDP data you wanted extracted from the
specified table.
}
\description{
This function will use the package \code{OECD} to extract GDP data for select
countries. This returns a tibble of all detailed GDP data which has been
harmonized the OECD for allow for country to country comparisons.
}
\examples{
\dontrun{

country_abbrevs <- get_oecd_country_abbrevs() \%>\%
 pull(.data$id)

GDP_dat <-
  get_oecd_gdp(
    variables = "QNA",
    countries = country_abbrevs,
    start_time = 2006,
    end_time = 2022,
    remove_abbrev_cols = TRUE
  )

}
}
