% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_data_helpers.R
\name{get_oecd_GDP_local}
\alias{get_oecd_GDP_local}
\title{This function will return a tibble containing country and names and
abbreviations used in the OECD database. The other OECD get functions
require country abbreviations to be provided as an input.}
\usage{
get_oecd_GDP_local()
}
\value{
(tibble) A tibble containing all country abbreviations and names.
}
\description{
This function draws in a local copy of the data set for speed. Update the
function for updated local copies of the data.
}
\examples{
\dontrun{

dat <- get_oecd_GDP_local()

}
}
