% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD_data_helpers.R
\name{get_oecd_labour_force_live}
\alias{get_oecd_labour_force_live}
\title{This function extracts relevant Employment data from the OECD using
"DP_LIVE" table. This function will return the following:}
\usage{
get_oecd_labour_force_live(
  variables = "DP_LIVE",
  countries = c("CHN", "USA", "GBR", "OECD", "G-7", "EU27_2020", "JPN", "ITA", "AUS",
    "CAN", "FRA", "DEU", "NZL", "NLD", "KOR", "IDN", "CHE"),
  start_time = 2020,
  end_time = 2022,
  remove_abbrev_cols = TRUE
)
}
\arguments{
\item{variables}{(character; "DP_LIVE") This variables which database table the
function will extract data from. The default is "DP_LIVE" which is the
the data table where labour force data is kept.}

\item{countries}{(character) Country abbreviations used by the function to
determine which countries will be extracted. Use \code{get_oecd_country_abbrevs}
to get the different abbreviations and associated country names.}

\item{start_time}{(numeric) The earliest year you want to go back to. The
function will attempt to go back as far as it can for the countries chosen.}

\item{end_time}{(numeric) Latest date you wish to extract data for.}

\item{remove_abbrev_cols}{(Boolean; TRUE) Will remove unnecessary abbreviation
columns if set to TRUE.}
}
\value{
(tibble)
}
\description{
G_EMP: Employment
EMP: Employment Rate
LF: Labour force
G_UNEMP: Unemployment
HUR: Unemployment rate
LTUNEMP: Long-term unemployment rate
}
\examples{
\dontrun{

dat <- get_oecd_labour_force_live(
countries = c("USA"),
start_time = 2020,
end_time = 2022
)

}
}
