% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australian_trade.R
\name{get_trade_DFAT_monthly}
\alias{get_trade_DFAT_monthly}
\title{This function obtains a local copy of DFAT Australian monthly trade data by
country and type of trade good.}
\usage{
get_trade_DFAT_monthly()
}
\value{
(tibble) Returns detailed monthly trade data for Australia.
}
\description{
This function obtains a local copy of DFAT Australian monthly trade data by
country and type of trade good.
}
\examples{
\dontrun{

dat <- get_trade_DFAT_monthly()

}
}
