% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australian_trade.R
\name{get_trade_DFAT_yearly}
\alias{get_trade_DFAT_yearly}
\title{This function obtains a local copy of DFAT Australian yearly trade data by
country and type of trade good.}
\usage{
get_trade_DFAT_yearly()
}
\value{
(tibble) Returns detailed yearly trade data for Australia.
}
\description{
This function obtains a local copy of DFAT Australian yearly trade data by
country and type of trade good.
}
\examples{
\dontrun {

dat <- get_trade_DFAT_yearly()

}
}
