% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_finance.R
\name{get_yahoo_finance}
\alias{get_yahoo_finance}
\title{This function will extract data from Yahoo finance based on the asset symbol
provided. The user specifies the start date and end date either as a character
or a date variable type. The function will return data for the asset between
the date ranges.}
\usage{
get_yahoo_finance(
  asset = "GBPUSD",
  start_date = "2008-01-01",
  end_date = lubridate::today()
)
}
\arguments{
\item{asset}{(character; "GBPUSD") Character value for the symbol of the
asset the user wishes to exrtact.}

\item{start_date}{(Character or date; "2008-01-01")}

\item{end_date}{(Character or date; lubridate::today())}
}
\value{
(tibble) Returns data extracted from Yahoo finance.
}
\description{
This function will extract data from Yahoo finance based on the asset symbol
provided. The user specifies the start date and end date either as a character
or a date variable type. The function will return data for the asset between
the date ranges.
}
\examples{
\dontrun{

asset <- "AUDUSD"
dat <- get_yahoo_finance(asset = "AUDUSD")

}
}
