% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoL_functions.R
\name{make_bls_query}
\alias{make_bls_query}
\title{This function will retrieve data based on the provided seriesid, start year
and end year input variables.}
\usage{
make_bls_query(
  url = "https://api.bls.gov/publicAPI/v1/timeseries/data/",
  seriesid = "CES0000000001",
  startyear = "2010",
  endyear = "2020"
)
}
\arguments{
\item{endyear}{}
}
\description{
This function will retrieve data based on the provided seriesid, start year
and end year input variables.
}
