% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{months_of_year}
\alias{months_of_year}
\title{Return a vector of months of the year}
\usage{
months_of_year(abbrev = FALSE)
}
\arguments{
\item{abbrev}{(Boolean; FALSE) Abbreviates the months if set to TRUE}
}
\description{
Return a vector of months of the year
}
\examples{
\dontrun{
days_week <- days_of_week()
}
}
