% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs.R
\name{get_abs_region_allocation}
\alias{get_abs_region_allocation}
\title{This function plucks concordance files from the ABS website and reads it into
your global environment.}
\usage{
get_abs_region_allocation(geo_type = "SA2")
}
\arguments{
\item{geo_type}{(character; SA2) Geo type, set to POA for postcodes and LGA
for LGA.}
}
\value{
(tibble)
}
\description{
This function plucks concordance files from the ABS website and reads it into
your global environment.
}
\examples{
\dontrun{

post_code_concordances <- get_abs_region_allocation("POA")

}
}
