% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niksmacrohelpers-package.R
\docType{package}
\name{niksmacrohelpers-package}
\alias{niksmacrohelpers}
\alias{niksmacrohelpers-package}
\title{niksmacrohelpers: Provides Users Free Macroeconomic and Financial Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nikhil Chandra \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
